Fixed Contact Form
Now messages that are submitted on the website get emailed directly to me! Before, nothing was connected to the form, so it did nothing. Used FormSpree to achieve, since this is a GitHubSite and therefore difficult/impossible to add a real back end to. I'm thinking keeping this site strictly front-end, and then I will create a real backend for my coffee-predictor project instead.